refactor: Change width parameter type to double in set_raydir function for improved precision

+++ project/src/renderer/set_raydir.c
@@ -12,9 +12,9 @@
 
 #include "cub3d.h"
 
-void	set_raydir(t_game *game, t_render *args, int width, int x)
+void	set_raydir(t_game *game, t_render *args, double width, double x)
 {
-	int	camerax;
+	double	camerax;
 
 	camerax = ((2 * x) / width) - 1; // width ekran boyutu olacak. 
 	args->ray_dirx = game->player.dir_x + args->plane_x * camerax;
